Extend mutil parameter configuration
